+++ RecruitmentSite/RecruitmentSite/Domain/Entities/CandidateExperience.cs
@@ -10,7 +10,7 @@ namespace RecruitmentSite.Domain.Entities
         public required string Company { get; set; }
         public required string Job { get; set; }
         public required string Description { get; set; }
-        public decimal Salary { get; set; }
+        public int Salary { get; set; }
         public DateTime BeginDate { get; set; }
         public DateTime? EndDate { get; set; }
         public DateTime InsertDate { get; set; }
